% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_setup.R
\name{run_setup}
\alias{run_setup}
\title{Load UCUM into Postgres}
\usage{
run_setup(
  conn,
  path_to_excel,
  schema = "ucum",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  log_schema = "public",
  log_table_name = "setup_ucum_log"
)
}
\arguments{
\item{conn}{Connection to a Postgres database.}

\item{schema}{Target schema. Default: 'ucum'.}

\item{log_schema}{Schema for the table that logs the process, Default: 'public'}

\item{log_table_name}{Name of log table, Default: 'setup_ucum_log'.}
}
\description{
The Excel file downloaded from
\url{https://ucum.org/trac/wiki/adoption/common} is loaded into Postgres.The
file release dates are taken along with the table names and row
counts are logged to a separate table after the loading is completed.
}
